feat: add proto_srcs to ts_proto_library (#152)
https://github.com/aspect-build/rules_ts/blob/main/docs/proto.md#ts_proto_library
says

copy_files | whether to copy the resulting .d.ts files back to the
source tree, for the editor to locate them. | True

This means gazelle currently trips users over
https://github.com/aspect-build/rules_ts/blob/main/ts/proto.bzl#L153

We could set copy_files=False by default but that's not Aspect's
recommendation since the editor isn't happy.
By setting proto_srcs we are able to copy outputs back to the source
tree.

Note that the copy is in a macro, so we can't read providers. Even if we
could, `ts_proto_library` rule cannot see the list of files, even though
`ProtoInfo.direct_sources` exists, because it points to
`_virtual_imports` when proto_library decides that path munging is
required (under `strip_import_prefix` for example)